Fix movement for Push Block + other additions
Rework the function that moves the block by calling a different method.
Added the "legacy" option for backwards compatibility.

This fixes a bug where the block travels further when pushed in one
direction compared to the other, i.e. further to the left than to the right,
further upwards than downwards. This also makes the block fall at the
same speed as a Falling Block.

As this fix is only effective as long as "legacy" is off, it should not affect
any Push Blocks that have already been placed.

Other additions include:
 - support for custom textures
 - small tweaks to debris and particle generation